kvm/riscv: handle some SBI extensions in userspace
KVM currently ignores what QEMU wants to have in registers on
sbi_hart_start; fix that by implementing HSM in QEMU.

We also need to implement BASE, because the guest wouldn't see HSM
otherwise.

Most of the code is fluff we need to properly capture SBI in the machine
model.  This should be properly split into many patches, but it's just a
preview for the kernel at this point.

DO NOT UPSTREAM.